# loops/testing/runs/001-quickflow/backend-dev/phase-05/attempt-1/C10-api-docs.json
{"openapi":"3.1.0","info":{"title":"QuickFlow API","version":"0.1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/tasks/{id}":{"get":{"tags":["tasks"],"operationId":"getTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}},"put":{"tags":["tasks"],"summary":"Replace the editable fields (status DONE sets completedAt, leaving DONE clears it)","operationId":"updateTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}},"delete":{"tags":["tasks"],"summary":"Delete permanently (BR-14)","operationId":"deleteTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/learning-cards/{id}":{"get":{"tags":["learning"],"operationId":"getLearningCard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearningCard"}}}}}},"put":{"tags":["learning"],"operationId":"updateLearningCard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearningCardUpdate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearningCard"}}}}}},"delete":{"tags":["learning"],"summary":"Delete the card, its milestones and notes permanently (BR-14)","operationId":"deleteLearningCard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/learning-cards/{id}/milestones/{milestoneId}":{"put":{"tags":["learning"],"summary":"Edit title/target date and set done or not done","operationId":"updateMilestone","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"milestoneId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MilestoneUpdate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Milestone"}}}}}},"delete":{"tags":["learning"],"operationId":"deleteMilestone","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"milestoneId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/habits/{id}":{"get":{"tags":["habits"],"operationId":"getHabit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Habit"}}}}}},"put":{"tags":["habits"],"operationId":"updateHabit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitWrite"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Habit"}}}}}},"delete":{"tags":["habits"],"summary":"Delete the habit and its completions permanently (BR-14)","operationId":"deleteHabit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/tasks":{"get":{"tags":["tasks"],"summary":"List tasks (search, filters, sort; archived excluded by default, BR-4)","operationId":"listTasks","parameters":[{"name":"q","in":"query","description":"Case-insensitive substring of the title","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"$ref":"#/components/schemas/TaskStatus"}},{"name":"priority","in":"query","required":false,"schema":{"$ref":"#/components/schemas/TaskPriority"}},{"name":"dueFrom","in":"query","description":"Inclusive","required":false,"schema":{"type":"string","format":"date"}},{"name":"dueTo","in":"query","description":"Inclusive","required":false,"schema":{"type":"string","format":"date"}},{"name":"archived","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"createdAt","enum":["createdAt","dueDate"]}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc","enum":["asc","desc"]}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}},"post":{"tags":["tasks"],"operationId":"createTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreate"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/api/tasks/{id}/restore":{"post":{"tags":["tasks"],"operationId":"restoreTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/api/tasks/{id}/complete":{"post":{"tags":["tasks"],"summary":"Mark completed (status DONE, BR-5)","operationId":"completeTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/api/tasks/{id}/archive":{"post":{"tags":["tasks"],"operationId":"archiveTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/api/learning-cards":{"get":{"tags":["learning"],"summary":"Cards with their milestones and notes, newest first","operationId":"listLearningCards","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LearningCard"}}}}}}},"post":{"tags":["learning"],"operationId":"createLearningCard","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearningCardCreate"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearningCard"}}}}}}},"/api/learning-cards/{id}/notes":{"post":{"tags":["learning"],"operationId":"addNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteCreate"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Note"}}}}}}},"/api/learning-cards/{id}/milestones":{"post":{"tags":["learning"],"operationId":"addMilestone","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MilestoneCreate"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Milestone"}}}}}}},"/api/habits":{"get":{"tags":["habits"],"summary":"Habits with progress, oldest first","operationId":"listHabits","parameters":[{"name":"active","in":"query","description":"Omit for all habits","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Habit"}}}}}}},"post":{"tags":["habits"],"operationId":"createHabit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitWrite"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Habit"}}}}}}},"/api/habits/{id}/deactivate":{"post":{"tags":["habits"],"operationId":"deactivateHabit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Habit"}}}}}}},"/api/habits/{id}/completions":{"get":{"tags":["habits"],"summary":"Completions, newest date first","operationId":"listHabitCompletions","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HabitCompletion"}}}}}}},"post":{"tags":["habits"],"summary":"Record a completion for a date (default today); one per habit and date (BR-7)","operationId":"completeHabit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitCompletionCreate"}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitCompletion"}}}}}}},"/api/habits/{id}/activate":{"post":{"tags":["habits"],"operationId":"activateHabit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Habit"}}}}}}},"/api/tasks/overdue":{"get":{"tags":["tasks"],"summary":"Tasks due before today, not Done, not archived (FR-01.7)","operationId":"listOverdueTasks","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}}},"/api/app-info":{"get":{"tags":["info"],"summary":"App time zone and the server's current time","operationId":"getAppInfo","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppInfo"}}}}}}},"/api/learning-cards/{id}/notes/{noteId}":{"delete":{"tags":["learning"],"operationId":"deleteNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"noteId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/habits/{id}/completions/{date}":{"delete":{"tags":["habits"],"operationId":"undoHabitCompletion","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"TaskPriority":{"type":"string","enum":["LOW","MEDIUM","HIGH"]},"TaskStatus":{"type":"string","enum":["TODO","IN_PROGRESS","DONE"]},"TaskUpdate":{"type":"object","properties":{"title":{"type":"string","maxLength":200,"minLength":1},"description":{"type":["string","null"],"maxLength":2000,"minLength":0},"status":{"$ref":"#/components/schemas/TaskStatus"},"priority":{"$ref":"#/components/schemas/TaskPriority"},"dueDate":{"type":["string","null"],"format":"date"}},"required":["priority","status","title"]},"Task":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":["string","null"]},"status":{"$ref":"#/components/schemas/TaskStatus"},"priority":{"$ref":"#/components/schemas/TaskPriority"},"dueDate":{"type":["string","null"],"format":"date"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"completedAt":{"type":["string","null"],"format":"date-time"},"archived":{"type":"boolean"},"overdue":{"type":"boolean","description":"Computed on read (FR-01.7)"}},"required":["archived","createdAt","id","overdue","priority","status","title","updatedAt"]},"LearningCardUpdate":{"type":"object","properties":{"title":{"type":"string","maxLength":200,"minLength":1},"description":{"type":["string","null"],"maxLength":2000,"minLength":0},"status":{"$ref":"#/components/schemas/LearningStatus"}},"required":["status","title"]},"LearningStatus":{"type":"string","enum":["NOT_STARTED","IN_PROGRESS","COMPLETED"]},"LearningCard":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":["string","null"]},"status":{"$ref":"#/components/schemas/LearningStatus"},"createdAt":{"type":"string","format":"date-time"},"milestones":{"type":"array","items":{"$ref":"#/components/schemas/Milestone"}},"notes":{"type":"array","items":{"$ref":"#/components/schemas/Note"}},"milestonesDone":{"type":"integer","format":"int32"},"milestonesTotal":{"type":"integer","format":"int32"}},"required":["createdAt","id","milestones","milestonesDone","milestonesTotal","notes","status","title"]},"Milestone":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cardId":{"type":"integer","format":"int64"},"title":{"type":"string"},"done":{"type":"boolean"},"targetDate":{"type":["string","null"],"format":"date"}},"required":["cardId","done","id","title"]},"Note":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cardId":{"type":"integer","format":"int64"},"text":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}},"required":["cardId","createdAt","id","text"]},"MilestoneUpdate":{"type":"object","properties":{"title":{"type":"string","maxLength":200,"minLength":1},"targetDate":{"type":["string","null"],"format":"date"},"done":{"type":"boolean"}},"required":["done","title"]},"HabitFrequency":{"type":"string","enum":["DAILY","WEEKLY"]},"HabitWrite":{"type":"object","properties":{"name":{"type":"string","maxLength":150,"minLength":1},"description":{"type":["string","null"],"maxLength":2000,"minLength":0},"frequency":{"$ref":"#/components/schemas/HabitFrequency"}},"required":["frequency","name"]},"Habit":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":["string","null"]},"frequency":{"$ref":"#/components/schemas/HabitFrequency"},"createdAt":{"type":"string","format":"date-time"},"active":{"type":"boolean"},"completedToday":{"type":"boolean"},"doneForCurrentPeriod":{"type":"boolean","description":"Today (DAILY) or current Monday-Sunday week (WEEKLY)"},"currentStreak":{"type":"integer","format":"int32","description":"Consecutive periods with a completion","minimum":0}},"required":["active","completedToday","createdAt","currentStreak","doneForCurrentPeriod","frequency","id","name"]},"TaskCreate":{"type":"object","properties":{"title":{"type":"string","maxLength":200,"minLength":1},"description":{"type":["string","null"],"maxLength":2000,"minLength":0},"status":{"$ref":"#/components/schemas/TaskStatus","description":"Default TODO"},"priority":{"$ref":"#/components/schemas/TaskPriority","description":"Default MEDIUM"},"dueDate":{"type":["string","null"],"format":"date"}},"required":["title"]},"LearningCardCreate":{"type":"object","properties":{"title":{"type":"string","maxLength":200,"minLength":1},"description":{"type":["string","null"],"maxLength":2000,"minLength":0}},"required":["title"]},"NoteCreate":{"type":"object","properties":{"text":{"type":"string","maxLength":5000,"minLength":1}},"required":["text"]},"MilestoneCreate":{"type":"object","properties":{"title":{"type":"string","maxLength":200,"minLength":1},"targetDate":{"type":["string","null"],"format":"date"}},"required":["title"]},"HabitCompletionCreate":{"type":"object","properties":{"date":{"type":["string","null"],"format":"date","description":"Default today; a future date is rejected"}}},"HabitCompletion":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"habitId":{"type":"integer","format":"int64"},"completionDate":{"type":"string","format":"date"},"createdAt":{"type":"string","format":"date-time"}},"required":["completionDate","createdAt","habitId","id"]},"AppInfo":{"type":"object","properties":{"timeZone":{"type":"string","examples":["Africa/Cairo"]},"now":{"type":"string","format":"date-time"}},"required":["now","timeZone"]}}}}